Replace PhoneAudioDeviceManager with broader AudioDeviceManager
